+++ src/SaveHandler/Cache.php
@@ -104,7 +104,12 @@ class Cache implements SaveHandlerInterface
      */
     public function destroy($id)
     {
-        return $this->getCacheStorage()->removeItem($id);
+        $this->getCacheStorage()->getItem($id, $exists);
+        if (! (bool) $exists) {
+            return true;
+        }
+
+        return (bool) $this->getCacheStorage()->removeItem($id);
     }
 
     /**